sliedrの汎用化テスト
